+++ lib/src/main/java/com/cloudbees/groovy/cps/impl/CpsCallableInvocation.java
@@ -3,7 +3,9 @@ package com.cloudbees.groovy.cps.impl;
 import com.cloudbees.groovy.cps.Block;
 import com.cloudbees.groovy.cps.Continuation;
 import com.cloudbees.groovy.cps.Env;
+import com.cloudbees.groovy.cps.Logging;
 import com.cloudbees.groovy.cps.Next;
+import java.io.PrintStream;
 
 import java.util.List;
 
@@ -47,6 +49,21 @@ public class CpsCallableInvocation extends Error/*not really an error but we wan
         return call.invoke(caller, loc, receiver,arguments,k);
     }
 
+    public Next invoke(String expectedMethodName, Env caller, SourceLocation loc, Continuation k) {
+        if (isMismatch(expectedMethodName, methodName)) {
+            PrintStream ps = Logging.current();
+            if (ps != null) {
+                ps.println(mismatchMessage(expectedMethodName, methodName));
+            }
+        }
+        return invoke(caller, loc, k);
+    }
+    
+    public static String mismatchMessage(String expectedMethodName, String actualMethodName) {
+        // TODO reference something like https://jenkins.io/redirects/pipeline-cps-method-mismatches/ sending you to a wiki page with commonly attempted idioms and the working equivalents
+        return "expected to call " + expectedMethodName + " but wound up catching " + actualMethodName;
+    }
+
     /**
      * Creates a {@link Block} that performs this invocation and pass the result to the given {@link Continuation}.
      */
@@ -66,6 +83,21 @@ public class CpsCallableInvocation extends Error/*not really an error but we wan
     @Override
     public String toString() {
         return "CpsCallableInvocation{methodName=" + methodName + ", call=" + call + ", receiver=" + receiver + ", arguments=" + arguments + '}';
+    }
+
+    /** @see <a href="https://issues.jenkins-ci.org/browse/JENKINS-31314">JENKINS-31314</a> */
+    private static boolean isMismatch(String expected, String caught) {
+        if (expected.equals(caught)) {
+            return false;
+        }
+        if (expected.startsWith("$")) {
+            // see TODO comment in Translator w.r.t. overloadsResolved
+            return false;
+        }
+        if (expected.equals("evaluate") && caught.equals("run")) {
+            return false;
+        }
+        return true;
     }
 
 }
